rewrite page table structure

+++ src/kernel/src/arch/x86_64/registers/control/cr4.rs
@@ -1,7 +1,7 @@
 bitflags! {
     #[repr(transparent)]
     #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-    pub struct CR4Flags: u64 {
+    pub struct CR4Flags: usize {
         const VME           = 1 << 0;
         const PVI           = 1 << 1;
         const TSD           = 1 << 2;
@@ -33,7 +33,7 @@ pub struct CR4;
 impl CR4 {
     #[inline]
     pub fn read() -> CR4Flags {
-        let value: u64;
+        let value: usize;
 
         // Safety: Reading into a register has no side effects.
         unsafe {
